+++ installation/routines/update_raspi_os.sh
@@ -1,9 +1,11 @@
 #!/usr/bin/env bash
 
 update_raspi_os() {
-  echo "Updating Raspberry Pi OS" | tee /dev/fd/3
+    if [ "$UPDATE_RASPI_OS" == true ] ; then
+        echo "Updating Raspberry Pi OS" | tee /dev/fd/3
 
-  sudo apt-get -qq -y update && sudo apt-get -qq -y full-upgrade && sudo apt-get -qq -y autoremove || exit_on_error "Failed to Update Raspberry Pi OS"
+        sudo apt-get -qq -y update && sudo apt-get -qq -y full-upgrade && sudo apt-get -qq -y autoremove || exit_on_error "Failed to Update Raspberry Pi OS"
 
-  echo "DONE: update_raspi_os"
+        echo "DONE: update_raspi_os"
+    fi
 }
